+++ common/dircmp_rcs_scanfile.c
@@ -1,5 +1,5 @@
 /*-
- * Copyright (c) 2000-2005 MAEKAWA Masahide <maekawa@cvsync.org>
+ * Copyright (c) 2000-2013 MAEKAWA Masahide <maekawa@cvsync.org>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -172,7 +172,7 @@ dircmp_rcs_scanfile(struct dircmp_args *dca)
 				break;
 			}
 		}
-		namelen = attr->a_namelen - (name - sv_name);
+		namelen = attr->a_namelen - (size_t)(name - sv_name);
 
 		if (namelen == cap->ca_namelen) {
 			rv = memcmp(name, cap->ca_name, namelen);
