+++ common/filescan.c
@@ -1,5 +1,5 @@
 /*-
- * Copyright (c) 2000-2005 MAEKAWA Masahide <maekawa@cvsync.org>
+ * Copyright (c) 2000-2013 MAEKAWA Masahide <maekawa@cvsync.org>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -269,8 +269,8 @@ filescan_start(struct filescan_args *fsa, const char *name, const char *relname)
 
 	SetWord(cmd, len - 2);
 	cmd[2] = FILECMP_START;
-	cmd[3] = namelen;
-	cmd[4] = relnamelen;
+	cmd[3] = (uint8_t)namelen;
+	cmd[4] = (uint8_t)relnamelen;
 	if (!mux_send(fsa->fsa_mux, MUX_FILECMP, cmd, 5))
 		return (false);
 	if (!mux_send(fsa->fsa_mux, MUX_FILECMP, name, namelen))
